First pass at always recalculating the dimensions
Also adds an optional timer to handle the cases where
you don't get notification of size changes.  The code
checks the previous size to see if it's changed and
avoids re-rendering if the size is the same, which
keeps the check fast.

Example:
I have a case where clicking on a graph causes it to
expand taller, which may result in a vertical scrollbar
being added on the right side.  This makes the view a
bit narrower, and my graph needs to shrink.  Without
the timer, it doesn't resize.  With the timer, it does.