BREAKING: measure parentNode instead of wrapper div
We probably should have done this before. Instead of measuring the dimensions of the wrapper `div` we reach up the DOM tree and measure the parent node. This should avoid some problems people have been having with rendering `0`. It might change behaviour in some cases though so this probably needs to be a major version bump.